pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "dbr-logs"
version = "0.1.0"
description = "Fetch and display Databricks job logs from Unity Catalog Volumes"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "databricks-sdk>=0.20.0",
]

[project.scripts]
dbr-logs = "dbr_logs.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/dbr_logs"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.14",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
